This is used for `MulMod()` so `big.Mul()` is not needed any more (we still need `big.Mod()`).
This makes `small/uint256-6` case slower because we also removed the shortcut for `128x128` case.
The same procedure can be used to implement `Mul()`, but when `umul()` is used directly the final result is slower - optimizer is not clever enough to notice that the top 4 words in `umul()` are not used in the end.
Inner loop unroll removes any bounds checking.
```
name                     old time/op  new time/op  delta
_MulMod/large/big-6       384ns ± 1%   382ns ± 0%   -0.47%  (p=0.048 n=5+5)
_MulMod/large/uint256-6   802ns ± 0%   721ns ± 0%  -10.05%  (p=0.000 n=5+4)
_MulMod/small/big-6      93.9ns ± 0%  93.6ns ± 0%     ~     (p=0.103 n=5+5)
_MulMod/small/uint256-6  33.5ns ± 0%  41.1ns ± 0%  +22.69%  (p=0.016 n=4+5)
```