@holiman 2020-03-14
Nice!
I pushed a commit on top, which adds a new conversion function from `*big.Int` to `uint256.Int`. If I apply that :
```diff
diff --git a/uint256.go b/uint256.go
index 63f3d4a..4ca0df9 100644
--- a/uint256.go
+++ b/uint256.go
@@ -618,7 +618,7 @@ func (z *Int) MulMod(x, y, m *Int) *Int {
        // if they fit within 256 bytes. But for now just wrap big.Int instead
        bp := new(big.Int)
        bp.SetBytes(pbytes[:])
-       z.SetFromBig(bp.Mod(bp, m.ToBig()))
+       z.SetFromBigFast(bp.Mod(bp, m.ToBig()))
        return z
 }
```
Then it's possible to further improve the worst case with 20% 
```
name                     old time/op    new time/op    delta
_MulMod/large/big-6         531ns ±21%     539ns ±18%     ~     (p=0.421 n=5+5)
_MulMod/large/uint256-6    1.08µs ±12%    0.86µs ± 0%  -20.57%  (p=0.016 n=5+4)
_MulMod/small/big-6         129ns ±17%     125ns ± 4%     ~     (p=0.849 n=5+5)
_MulMod/small/uint256-6    51.4ns ± 0%    52.3ns ± 5%     ~     (p=1.000 n=5+5)

name                     old alloc/op   new alloc/op   delta
_MulMod/large/big-6          176B ± 0%      176B ± 0%     ~     (all equal)
_MulMod/large/uint256-6      336B ± 0%      304B ± 0%   -9.52%  (p=0.008 n=5+5)
_MulMod/small/big-6         56.0B ± 0%     56.0B ± 0%     ~     (all equal)
_MulMod/small/uint256-6     0.00B          0.00B          ~     (all equal)

name                     old allocs/op  new allocs/op  delta
_MulMod/large/big-6          2.00 ± 0%      2.00 ± 0%     ~     (all equal)
_MulMod/large/uint256-6      6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.008 n=5+5)
_MulMod/small/big-6          2.00 ± 0%      2.00 ± 0%     ~     (all equal)
_MulMod/small/uint256-6      0.00           0.00          ~     (all equal)
```

I didn't commit actually using that function, because it's not yet complete - I haven't implemented support for 32-bit platforms. 

Tangent: is it possible/easy to enable 32-bit platform testing on CircleCI?