@chfast 2020-03-14
Hey, I moved your commit to `conversion` branch. I'd like to take a look there later. In the end the conversion from/to `math.big` will not affect performance in `MulMod` as we will have our own division procedure.