@chfast 2020-03-14
> Tangent: is it possible/easy to enable 32-bit platform testing on CircleCI?

Done in https://github.com/holiman/uint256/pull/31, but currently fails in a conversion test.